MAINT: Move exceptions from _definitions to exceptions file (#1001)
