+++ tests/test_units/test_objectdataprovider_class.py
@@ -7,8 +7,9 @@ import pytest
 import yaml
 
 from fmu import dataio
-from fmu.dataio._definitions import ConfigurationError, ValidFormats
+from fmu.dataio._definitions import ValidFormats
 from fmu.dataio._models.fmu_results.specification import FaultRoomSurfaceSpecification
+from fmu.dataio.exceptions import ConfigurationError
 from fmu.dataio.providers.objectdata._faultroom import FaultRoomSurfaceProvider
 from fmu.dataio.providers.objectdata._provider import (
     objectdata_provider_factory,
